+++ Private/DFCatalog.Scoop.ps1
@@ -368,7 +368,7 @@ function Get-DFCatalogScoopDetail {
         -Extra $extra
 }
 
-if (-not $script:DFCatalogProviders) { $script:DFCatalogProviders = @{} }
+if (-not (Get-Variable -Name DFCatalogProviders -Scope Script -ErrorAction Ignore)) { $script:DFCatalogProviders = @{} }
 $script:DFCatalogProviders['scoop'] = @{
     Name         = 'scoop'
     Kind         = 'snapshot'